Update: Refactor posts form
