+++ src/entities/posts/index.ts
@@ -1,0 +1,2 @@
+export { usePostForm, PostSchema } from './model/';
+export type { PostSchemaType } from './model/';
