+++ src/entities/posts/model/index.ts
@@ -1,1 +1,4 @@
 export { PostStatus, type PostStatusType, PostStatusList } from './types';
+export { usePostForm } from '../api/use-post-form';
+export { PostSchema } from './schema';
+export type { PostSchema as PostSchemaType } from './schema';
