+++ src/entities/posts/model/schema.ts
