+++ src/features/admin/add-post-form/api/index.ts
@@ -1,2 +1,1 @@
 export { useFetchLabels, useSubmitPost } from './queries';
-export { PostSchema } from './schema';
